wtf happened
Someone is messing up with my files...Git moves randonly without my permissions, and randomly skipes commits....are someone plays with it...!? NOT FUNNY! Get yourself a girlfriend and get away from here, you jerk!

+++ php/uprofile.php
@@ -27,13 +27,12 @@ if (isset($_POST["uploadProfile"])) {
     $imgname = $profilePicture;
   }
 
-  if ($imgname != $profilePicture) {
+  if ($imgname != $profilePicture && $profilePicture != null) {
     $path = "../images/profiles/".$_SESSION["username"]."/".$profilePicture;
     chown($path,0777);
     unlink($path);
   }
   
-
 // Check if image file is a actual image or fake image
   if(isset($_POST["uploadProfile"])) {
     $check = getimagesize($_FILES["fileToUpload"]["tmp_name"]);
